+++ docs/guide/demos/wagmi.tsx
@@ -1,20 +1,31 @@
 import { WagmiConfig, createConfig, configureChains, mainnet } from 'wagmi';
 import { publicProvider } from 'wagmi/providers/public';
+import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
+import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
 import { WagmiWeb3ConfigProvider } from '@ant-design/web3-wagmi';
-import { NFTImage } from '@ant-design/web3';
+import { ConnectButton } from '@ant-design/web3';
 
-const { publicClient } = configureChains([mainnet], [publicProvider()]);
+const { publicClient, chains } = configureChains([mainnet], [publicProvider()]);
 
 const config = createConfig({
   autoConnect: true,
   publicClient,
+  connectors: [
+    new MetaMaskConnector(),
+    new WalletConnectConnector({
+      chains,
+      options: {
+        projectId: YOUR_WALLET_CONNET_PROJECT_ID,
+      },
+    }),
+  ],
 });
 
 export default () => {
   return (
     <WagmiConfig config={config}>
       <WagmiWeb3ConfigProvider>
-        <NFTImage address="0x79fcdef22feed20eddacbb2587640e45491b757f" tokenId={42} />
+        <ConnectButton />
       </WagmiWeb3ConfigProvider>
     </WagmiConfig>
   );
