+++ packages/common/src/index.ts
@@ -1,7 +1,6 @@
 export * from './types';
-export * from './wallets/index';
-export * from './rpc-providers';
 export * from './utils';
 export * from './web3-config-provider';
+export * from './wallets';
 
 export * as chains from './chains';
