+++ packages/common/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "baseUrl": "."
   },
-  "include": ["src", "global.d.ts"]
+  "include": ["src", "global.d.ts", "../ethereum/src/wallets"]
 }
