+++ packages/ethereum/src/index.ts
@@ -1,6 +1,7 @@
-export { ZANJsonRpcProvider } from '@ant-design/web3-common';
-export { MetaMaskProvider, WalletConnectProvider } from '@ant-design/web3-common';
 export { chains } from '@ant-design/web3-common';
 
+export * from './wallets';
+export * from './rpc-providers';
 export * from './ethereum-provider';
 export * from './universal-provider';
+export * from './types';
