+++ packages/ethereum/src/wallets/index.ts
@@ -1,0 +1,2 @@
+export * from './meta-mask';
+export * from './wallet-connect';
