+++ packages/wagmi/package.json
@@ -1,6 +1,8 @@
 {
   "name": "@ant-design/web3-wagmi",
   "version": "0.1.0-alpha.1",
+  "type": "module",
+  "main": "./dist/esm/index.js",
   "exports": {
     "import": "./dist/esm/index.js",
     "require": "./dist/lib/index.js",
