+++ packages/wagmi/src/global.d.ts
@@ -1,0 +1,3 @@
+declare interface Window {
+  ethereum?: any;
+}
