+++ packages/wagmi/src/wagmi-provider.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import { Web3ConfigProvider } from '@ant-design/web3-common';
+import { Web3ConfigProvider, Account, Wallet } from '@ant-design/web3-common';
+import { useAccount, useConnect } from 'wagmi';
+import { getWalletsByConnectors } from './wallets.js';
 
 export interface WagmiWeb3ConfigProviderProps {
   children?: React.ReactNode;
@@ -7,6 +9,41 @@ export interface WagmiWeb3ConfigProviderProps {
 
 export const WagmiWeb3ConfigProvider: React.FC<WagmiWeb3ConfigProviderProps> = (props) => {
   const { children } = props;
+  const { address } = useAccount();
+  const { connectors, connectAsync } = useConnect();
 
-  return <Web3ConfigProvider>{children}</Web3ConfigProvider>;
+  const accounts: Account[] = React.useMemo(() => {
+    if (!address) {
+      return [];
+    }
+    return [
+      {
+        address,
+      },
+    ];
+  }, [address]);
+
+  const wallets: Wallet[] = React.useMemo(() => {
+    return getWalletsByConnectors(connectors);
+  }, [connectors]);
+
+  return (
+    <Web3ConfigProvider
+      accounts={accounts}
+      wallets={wallets}
+      requestAccounts={async (wallet) => {
+        const connector = connectors.find((item) => item.name === wallet);
+        const { account } = await connectAsync({
+          connector,
+        });
+        return [
+          {
+            address: account,
+          },
+        ];
+      }}
+    >
+      {children}
+    </Web3ConfigProvider>
+  );
 };
